+++ src/components/Comparison/hooks/useComparisonVersions.js
@@ -1,21 +1,27 @@
 import { useEffect, useState, useCallback } from 'react';
+import { useSelector, useDispatch } from 'react-redux';
 import { getComparisonVersionsById } from 'services/backend/comparisons';
+import { setVersions } from 'slices/comparisonSlice';
 
 function useComparisonVersions({ cId }) {
     const [isLoadingVersions, setIsLoadingVersions] = useState(false);
-    const [versions, setVersions] = useState([]);
     const [hasNextVersion, setHasNextVersion] = useState(false);
+    const versions = useSelector(state => state.comparison.versions);
+    const dispatch = useDispatch();
 
-    const loadVersions = useCallback(id => {
-        if (id) {
-            setIsLoadingVersions(true);
-            getComparisonVersionsById(id).then(v => {
-                setVersions(v);
-                setIsLoadingVersions(false);
-                setHasNextVersion(!!(v?.length > 0 && v[0].id !== id));
-            });
-        }
-    }, []);
+    const loadVersions = useCallback(
+        id => {
+            if (id && versions.length === 0) {
+                setIsLoadingVersions(true);
+                getComparisonVersionsById(id).then(v => {
+                    dispatch(setVersions(v));
+                    setIsLoadingVersions(false);
+                    setHasNextVersion(!!(v?.length > 0 && v[0].id !== id));
+                });
+            }
+        },
+        [dispatch, versions.length],
+    );
 
     useEffect(() => {
         loadVersions(cId);